@@ -1,16 +1,21 @@
-namespace Community.Blazor.MapLibre.Models.Source;
+using System.Text.Json.Serialization;
+using Community.Blazor.MapLibre.Models.Feature;
+
+namespace Community.Blazor.MapLibre.Models.Sources;
 
 /// <summary>
 /// Represents a GeoJSON source. GeoJSON sources provide either inline GeoJSON data or a URL to a GeoJSON file.
 /// They can support clustering and other custom behaviors for point features.
 /// </summary>
-public class GeoJSONSource : SourceBase
+public class GeoJsonSource : ISource
 {
     /// <inheritdoc />
-    public override string Type => "geojson";
+    [JsonPropertyName("type")]
+    public string Type => "geojson";
 
     /// <summary>
     /// The GeoJSON data, either as an inline object or a URL to an external GeoJSON file. Required.
     /// </summary>
-    public object Data { get; set; } = null!;
+    [JsonPropertyName("data")]
+    public required IFeature Data { get; set; } = null!;
 }
