@@ -1,25 +1,31 @@
-namespace Community.Blazor.MapLibre.Models.Source;
+using System.Text.Json.Serialization;
+
+namespace Community.Blazor.MapLibre.Models.Sources;
 
 /// <summary>
 /// Represents a raster tile source. Raster sources provide tiled raster image data, typically used for basemaps.
 /// </summary>
-public class RasterTileSource : SourceBase
+public class RasterTileSource : ISource
 {
     /// <inheritdoc />
-    public override string Type => "raster";
+    [JsonPropertyName("type")]
+    public string Type => "raster";
 
     /// <summary>
     /// URL to a TileJSON resource providing metadata about this source. Optional.
     /// </summary>
+    [JsonPropertyName("url")]
     public string? Url { get; set; }
 
     /// <summary>
     /// An array of URLs to the raster tiles. URL patterns can use placeholders like `{z}`, `{x}`, and `{y}`. Optional.
     /// </summary>
+    [JsonPropertyName("tiles")]
     public List<string>? Tiles { get; set; }
 
     /// <summary>
     /// The bounding box for the source, specified as an array `[sw.lng, sw.lat, ne.lng, ne.lat]`. Optional.
     /// </summary>
+    [JsonPropertyName("bounds")]
     public double[]? Bounds { get; set; }
 }
